Renderer runtime: Switched to 64 bit world space position. For testing I made a local change inside "SceneResourceLoader.cpp" -> "nodeDeserialization()" after reading a node by adding an 100.000.000 offset to the node transform position. Before the 64 bit world space position even +10.000 created notable issues, now even with an insane offset I notice no difference.

+++ Example/Source/Examples/Private/Runtime/FirstScene/FreeCameraController.cpp
@@ -91,7 +91,7 @@ void FreeCameraController::onUpdate(float pastSecondsSinceLastFrame, bool inputE
 	{
 		// Get the current local transform
 		const RendererRuntime::Transform& transform = sceneNode->getTransform();
-		glm::vec3 newPosition = transform.position;
+		glm::dvec3 newPosition = transform.position;	// 64 bit world space position
 		glm::quat newRotation = transform.rotation;
 
 		// In case input is enabled, process input
@@ -125,7 +125,7 @@ void FreeCameraController::onUpdate(float pastSecondsSinceLastFrame, bool inputE
 					}
 				}
 
-				// Get the movement vector
+				// Get the movement vector, 32 bit is sufficient here
 				glm::vec3 movementVector = RendererRuntime::Math::VEC3_ZERO;
 				{
 					{ // Move forward/backward
